requirements.txt
flask
flask-cors>=4.0
python-dotenv
google-generativeai
requests
# Database and Migrations
Flask-SQLAlchemy>=3.1
Flask-Migrate>=4.0
psycopg2-binary>=2.9
# Auth and Security
Flask-JWT-Extended>=4.6
Flask-Limiter>=3.5
Flask-Talisman>=1.1
bcrypt
# Validation
marshmallow>=3.21
# AI and Data
sentence-transformers
scikit-learn==1.7.2
pandas
numpy
numexpr
joblib
rapidfuzz
# Other utilities
werkzeug
reportlab
pypdf2
